Refactor - Update client versions (#52)
