+++ client/package.json
@@ -8,27 +8,27 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "bootstrap-vue": "^2.16.0",
-    "core-js": "^3.6.4",
-    "jquery": "^3.5.1",
-    "moment-timezone": "^0.5.31",
-    "vue": "^2.6.12",
-    "vue-clipboard2": "^0.3.1",
-    "vue-code-highlight": "^0.7.5",
-    "vue-json-pretty": "^1.7.1",
-    "vue-moment": "^4.1.0",
-    "vue-router": "^3.4.3",
-    "vue2-filters": "^0.11.0"
+    "axios": "^1.0.0",
+    "bootstrap-vue": "^2.0.0",
+    "core-js": "^3.0.0",
+    "moment-timezone": "^0.5.39",
+    "vue": "^2.0.0",
+    "vue-clipboard2": "^0.3.3",
+    "vue-code-highlight": "^0.7.7",
+    "vue-json-pretty": "^1.0.0",
+    "vue-moment": "^4.0.0",
+    "vue-router": "^3.0.0",
+    "vue2-filters": "^0.14.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.3.0",
-    "@vue/cli-plugin-eslint": "~4.3.0",
-    "@vue/cli-service": "^4.5.13",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.12"
+    "@vue/cli-plugin-babel": "^5.0.0",
+    "@vue/cli-plugin-eslint": "^5.0.0",
+    "@vue/cli-service": "^5.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/eslint-parser": "^7.0.0",
+    "eslint": "^7.0.0",
+    "eslint-plugin-vue": "^8.0.0",
+    "vue-template-compiler": "^2.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -40,7 +40,7 @@
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
     "rules": {}
   },
