+++ client/src/components/index/NavigationRow.vue
