+++ client/src/main.js
@@ -1,8 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 import App from './App'
-import Login from '@/components/login/Login'
-import Index from '@/components/index/Index'
+import Login from '@/components/login/LoginContainer'
+import Index from '@/components/index/IndexContainer'
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import { BootstrapVue, IconsPlugin, BIcon, BIconArrowRepeat, BIconChevronDoubleUp, BIconChevronDoubleDown, BIconTrash, BIconInfo, BIconGear } from 'bootstrap-vue'
